feat: locks can now be used by single player for one min without fee (#69)

+++ server/db/revert/base_schema.sql
@@ -12,7 +12,7 @@ DROP FUNCTION IF EXISTS fish(player_id INTEGER);
 
 DROP FUNCTION IF EXISTS purchase_item(player_id INTEGER, item_id INTEGER);
 
-DROP FUNCTION IF EXISTS player_markers(player_id INTEGER, marker_type TEXT, marker_distance_limit INTEGER);
+DROP FUNCTION IF EXISTS player_markers(player_id INTEGER, marker_type TEXT, marker_distance_limit FLOAT);
 
 DROP VIEW IF EXISTS links_recursive;
